Added option for VE prior
User can now set the mu for VE rather than only logRR. The sigma input is on the scale of logRR, so as the mu for VE increases, the sigma will get tighter since it nears the boundary of 1. When setting the prior on only VE, it is implicitly assumed that logRR follows a normal prior, so now Gaussian mixture. Seems simpler this way.